+++ lib/monko-core/src/jsMain/kotlin/MonkoClientJs.kt
@@ -1,7 +1,7 @@
 package dev.petuska.monko.core
 
 import dev.petuska.monko.core.ext.MongoClient
-import dev.petuska.monko.core.ext.connect
+import dev.petuska.monko.core.ext.MongodbModule
 import dev.petuska.monko.core.util.noop
 import kotlinx.coroutines.await
 
@@ -10,7 +10,7 @@ public actual suspend fun MonkoClient(
   connectionStringJS: String?,
   connectionStringJVM: String?,
   connectionStringNative: String?
-): MonkoClient = MonkoClientJs(connect(connectionStringJS ?: connectionString).await())
+): MonkoClient = MongodbModule.MongoClient.connect(connectionStringJS ?: connectionString).await().let(::MonkoClientJs)
 
 internal class MonkoClientJs(override val source: MongoClient) : MonkoClient {
   override suspend fun database(name: String): MonkoDatabase = MonkoDatabaseJs(this, name, source.db(name))